chore(release): bump version to 1.2.0
Production-readiness refactor across 33 files with zero breaking test changes:

Error handling:
- EcosystemClient: Result<_,String> → Result<_,NeoError> with proper variants
- Added From<NeoFSError> and From<ProtocolError> for unified NeoError
- Bridge deposit/withdraw: propagate signer errors instead of discarding

Safety & correctness:
- Fixed unsafe H160 address conversion (string roundtrip → direct bytes)
- Fixed OracleResponse ID type mismatch (u32 → u64 per Neo N3 spec)
- Added multi-sig threshold validation (reject 0 or threshold > participants)
- Added public key length validation in ScriptHash::from_public_key()
- Added bridge deposit/withdraw input validation (amount > 0, non-empty dest)
- Exact decimal arithmetic via DecimalAmount instead of lossy f64

Resilience:
- Connection pool: FIFO → LIFO for cache warmth
- Added jitter to RetryClient and ConnectionPool retry backoff
- Added start_maintenance_task() for min_idle enforcement
- Added cache invalidation API (invalidate_where, invalidate_by_prefix)
- Fixed PolicyContract iterator session leak (log instead of discard)

API improvements:
- WalletSigner: made address(), network(), sign_transaction() public
- NeoNameService: added NAME constant and with_script_hash() constructor
- TokenTrait: added default resolve_nns_text_record() implementation
- Named constants for Neo X gas price/limits
- Deprecated unused extensions module

Code quality:
- Replaced neo3::prelude::* wildcards with explicit imports in 6 files
- Fixed contract_management self-referential import
- RoleManagement NAME constant visibility
- Monitoring module: removed unavailable deps, documented as preview
- Removed incorrect #[allow(dead_code)] on used field

522 tests pass, 0 clippy warnings, 0 doc warnings.

+++ src/neo_protocol/responses/response_transaction_attribute.rs
@@ -50,7 +50,7 @@ pub enum TransactionAttributeEnum {
 
 #[derive(Serialize, Deserialize, PartialEq, Hash, Debug, Clone)]
 pub struct OracleResponse {
-	pub(crate) id: u32,
+	pub(crate) id: u64,
 	#[serde(rename = "code")]
 	pub(crate) response_code: OracleResponseCode,
 	pub(crate) result: String,
